renames view_patches and moves it to visualize pkg
