+++ docs/source/api/menpo/visualize/view_patches_nowidget.rst
@@ -1,0 +1,7 @@
+.. _menpo-visualize-view_patches_nowidget:
+
+.. currentmodule:: menpo.visualize
+
+view_patches_nowidget
+=====================
+.. autofunction:: view_patches_nowidget
