+++ menpo/image/__init__.py
@@ -2,4 +2,3 @@ from .base import (Image, ImageBoundaryError,
                    convert_patches_list_to_single_array)
 from .boolean import BooleanImage
 from .masked import MaskedImage, OutOfMaskSampleError
-from .visualize_patches import view_patches, _create_patches_image
